refactor: extract issue comment submission logic
