+++ ui/src/components/CommentThread.tsx
@@ -10,7 +10,7 @@ import { useDialog } from "@/context/DialogContext";
 import { applyOrganizationPrefix, extractOrganizationPrefixFromPath, toOrganizationRelativePath } from "@/lib/organization-routes";
 import { PluginSlotOutlet } from "@/plugins/slots";
 import type { Agent, IssueComment } from "@rudderhq/shared";
-import { buildIssueMentionHref, extractAgentWakeMentionIds, parseShortRef } from "@rudderhq/shared";
+import { buildIssueMentionHref } from "@rudderhq/shared";
 import { Check, ChevronDown, Copy, Link2, MoreHorizontal, Paperclip, Pencil, TerminalSquare, Trash2 } from "lucide-react";
 import { memo, useCallback, useEffect, useMemo, useRef, useState, type CSSProperties, type ChangeEvent, type KeyboardEvent, type MouseEvent, type ReactNode, type RefObject } from "react";
 import { Link, useLocation, useNavigate } from "react-router-dom";
@@ -21,6 +21,7 @@ import { formatRunDurationLabel, formatRunTimingTitle, isRunTimingActive } from 
 import { formatDateTime, relativeTime } from "../lib/utils";
 import { AgentIdentity } from "./AgentAvatar";
 import { commentThreadTranscriptRuns, type LinkedRunItem } from "./CommentThread.runs";
+import { useCommentSubmit } from "./CommentThread.submit";
 import { Identity } from "./Identity";
 import type { MarkdownAgentMentionPreview, MarkdownLinkClickHandler } from "./MarkdownBody";
 import { MarkdownBody } from "./MarkdownBody";
@@ -81,25 +82,6 @@ export function shouldOfferReopen(issueStatus?: string) {
   return issueStatus === "done";
 }
 
-export function shouldConfirmUnmentionedComment(
-  markdown: string,
-  validAgentIds: ReadonlySet<string> = new Set(),
-  reopenWillWakeAgent = false,
-) {
-  if (reopenWillWakeAgent) return false;
-  return !extractAgentWakeMentionIds(markdown).some((agentRef) => {
-    if (validAgentIds.has(agentRef)) return true;
-    const shortRef = parseShortRef(agentRef);
-    if (shortRef?.kind !== "agent") return false;
-    let matches = 0;
-    for (const agentId of validAgentIds) {
-      if (agentId.replace(/-/g, "").toLowerCase().startsWith(shortRef.prefix)) matches += 1;
-      if (matches > 1) return false;
-    }
-    return matches === 1;
-  });
-}
-
 function loadDraft(draftKey: string): string {
   try {
     return localStorage.getItem(draftKey) ?? "";
@@ -133,14 +115,6 @@ function shouldForwardComposerFocus(target: EventTarget | null) {
     "[role='menuitem']",
     "[data-chat-composer-menu-item]",
   ].join(","));
-}
-
-function clearDraft(draftKey: string) {
-  try {
-    localStorage.removeItem(draftKey);
-  } catch {
-    // Ignore localStorage failures.
-  }
 }
 
 function asRecord(value: unknown): Record<string, unknown> | null {
@@ -1082,12 +1056,9 @@ export function CommentThread({
   fixedComposerTimelineScroll = true,
   timelineScrollElementRef,
 }: CommentThreadProps) {
-  const { confirm } = useDialog();
   const [body, setBody] = useState(() => draftKey ? loadDraft(draftKey) : "");
   const canReopen = shouldOfferReopen(issueStatus);
   const [reopen, setReopen] = useState(canReopen);
-  const [submitting, setSubmitting] = useState(false);
-  const [confirmingUnmentioned, setConfirmingUnmentioned] = useState(false);
   const [attaching, setAttaching] = useState(false);
   const [highlightCommentId, setHighlightCommentId] = useState<string | null>(null);
   const [runExpandedOverrides, setRunExpandedOverrides] = useState<Record<string, boolean>>({});
@@ -1106,8 +1077,6 @@ export function CommentThread({
   const pendingScrollRetryTimersRef = useRef<Array<ReturnType<typeof setTimeout>>>([]);
   const lastCommentScrollRef = useRef<{ container: HTMLElement | null; scrollTop: number | null } | null>(null);
   const pendingScrollCancelCleanupRef = useRef<(() => void) | null>(null);
-  const submissionInFlightRef = useRef(false);
-  const validAgentIds = useMemo(() => new Set(agentMap?.keys() ?? []), [agentMap]);
   const visibleComments = useMemo(() => comments.filter((comment) => !comment.deletedAt), [comments]);
   const currentIssueId = visibleComments[0]?.issueId ?? null;
 
@@ -1200,6 +1169,19 @@ export function CommentThread({
     setBody(nextBody);
     if (draftKey) saveDraft(draftKey, nextBody);
   }, [draftKey]);
+  const { canSubmit, handleSubmit, submitting } = useCommentSubmit({
+    agentMap,
+    body,
+    canReopen,
+    composerSurfaceRef,
+    draftKey,
+    editorRef,
+    onAdd,
+    reopen,
+    reopenWillWakeAgent,
+    setReopen,
+    updateBody,
+  });
 
   const agentMentions = useMemo<MarkdownAgentMentionPreview[]>(() => (
     mentions
@@ -1403,47 +1385,6 @@ export function CommentThread({
     setReserveHashScrollEndSpace(false);
   }, [location.hash, location.key, scrollToComment]);
 
-  async function handleSubmit() {
-    if (submissionInFlightRef.current) return;
-    const currentMarkdown = editorRef.current?.getMarkdown?.() ?? body;
-    const trimmed = currentMarkdown.trim();
-    if (!trimmed) return;
-    const reopenRequested = canReopen && reopen ? true : undefined;
-
-    submissionInFlightRef.current = true;
-    try {
-      if (shouldConfirmUnmentionedComment(
-        trimmed,
-        validAgentIds,
-        Boolean(reopenRequested && reopenWillWakeAgent),
-      )) {
-        setConfirmingUnmentioned(true);
-        const confirmed = await confirm({
-          title: "未 @ 任何 Agent",
-          description: "您未 @ 任何 Agent，是否确认直接发送评论？未 @ Agent 的评论不会触发 Agent，可能无法被及时处理。",
-          cancelLabel: "返回并 @ Agent",
-          confirmLabel: "直接发送",
-          restoreFocus: (confirmed) => {
-            if (confirmed) composerSurfaceRef.current?.focus({ preventScroll: true });
-            else editorRef.current?.focus();
-          },
-        });
-        setConfirmingUnmentioned(false);
-        if (!confirmed) return;
-      }
-
-      setSubmitting(true);
-      await onAdd(trimmed, reopenRequested);
-      updateBody("");
-      if (draftKey) clearDraft(draftKey);
-      setReopen(canReopen);
-    } finally {
-      setSubmitting(false);
-      setConfirmingUnmentioned(false);
-      submissionInFlightRef.current = false;
-    }
-  }
-
   async function handleAttachFile(evt: ChangeEvent<HTMLInputElement>) {
     const files = Array.from(evt.target.files ?? []);
     if (files.length === 0) return;
@@ -1475,7 +1416,6 @@ export function CommentThread({
     }
   }
 
-  const canSubmit = !submitting && !confirmingUnmentioned && !!body.trim();
   const focusComposerEditor = (event: MouseEvent<HTMLDivElement>) => {
     if (!shouldForwardComposerFocus(event.target)) return;
     event.preventDefault();
