+++ ui/src/pages/Chat.side-panel.tsx
@@ -618,8 +618,7 @@ function ChatIssueSidePanelView({
               comments={comments}
               orgId={issue.orgId}
               projectId={issue.projectId}
-              issueStatus={issue.status}
-              reopenWillWakeAgent={Boolean(issue.assigneeAgentId)}
+              issueStatus={issue.status} reopenWillWakeAgent={Boolean(issue.assigneeAgentId)}
               agentMap={agentMap}
               currentUserId={currentUserId}
               operatorDisplayName={operatorDisplayName}
