+++ ui/src/pages/IssueDetail.tsx
@@ -2611,8 +2611,7 @@ export function IssueDetail({ embeddedIssueId = null, embedded = false }: IssueD
           activityItems={issueActivityItems}
           orgId={issue.orgId}
           projectId={issue.projectId}
-          issueStatus={issue.status}
-          reopenWillWakeAgent={Boolean(issue.assigneeAgentId)}
+          issueStatus={issue.status} reopenWillWakeAgent={Boolean(issue.assigneeAgentId)}
           agentMap={agentMap}
           draftKey={`rudder:issue-comment-draft:${issue.id}`}
           mentions={mentionOptions}
